requirements.txt
google-api-python-client==1.5.2
httplib2==0.9.2
oauth2client==3.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
